fix(ci): fix FileUtil.listExperimentFiles and TimeUtilTest Long comparison
- FileUtil.listExperimentFiles: create directory if it doesn't exist before listing
- TimeUtilTest.testGetTimeSlotStartMillis: cast intervalMinutes to Long to avoid Integer vs Long comparison

+++ ble-feasibility-lab/src/test/kotlin/com/ghost/blelab/util/TimeUtilTest.kt
@@ -84,7 +84,7 @@ class TimeUtilTest {
         val start = TimeUtil.getTimeSlotStartMillis(slot, intervalMinutes)
         val end = TimeUtil.getTimeSlotEndMillis(slot, intervalMinutes)
         
-        assertEquals("Slot duration should be interval minutes", intervalMinutes * 60 * 1000, end - start + 1)
+        assertEquals("Slot duration should be interval minutes", intervalMinutes.toLong() * 60 * 1000, end - start + 1)
     }
 
     @Test
